refactor(bindings/python): split layers into a package for extensibility
Split the single layers.rs into a src/layers/ package with one module per
layer plus a mod.rs holding the shared PythonLayer trait, the base Layer, and
re-exports. The Python surface stays flat on opendal.layers, so all existing
imports are unchanged and stub generation needs no special handling.

Adding a layer later is just: drop src/layers/<name>.rs, register it in
mod.rs and lib.rs, and run just stub-gen.
